[saved objects] enable deletion of saved objects by type if configured (#6443)
* [saved objects] enable deletion of saved objects by type if configured

Adds the following settings:
```
migrations.delete.enabled
migrations.delete.types
```

`unknown` types already exist but the purpose of this type is for plugins
that are disabled. OpenSearch Dashboards gets confused when a plugin is not
defining a saved object type but the saved object exists. This can occur
when migrating from a non-OSD version and there exists non-compatiable
saved objects.

If OSD is failing to migrate an index because of a document, I can now
configure OSD to delete types of saved objects that I specified because
I know that these types should not be carried over.

resolves: https://github.com/opensearch-project/OpenSearch-Dashboards/issues/1040

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>

* address comments

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>

* Changeset file for PR #6443 created/updated

---------

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>
Co-authored-by: opensearch-changeset-bot[bot] <154024398+opensearch-changeset-bot[bot]@users.noreply.github.com>